refactor : 테스트 객체 생성을 담당하는 test fixture로 변경
