+++ src/main/java/com/project/shop/item/controller/ItemController.java
@@ -3,7 +3,7 @@ package com.project.shop.item.controller;
 
 import com.project.shop.item.dto.response.ItemListResponse;
 import com.project.shop.item.dto.response.ItemResponse;
-import com.project.shop.item.service.ItemServiceImpl;
+import com.project.shop.item.service.ItemService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -17,20 +17,20 @@ import java.util.List;
 @RestController
 public class ItemController {
 
-    private final ItemServiceImpl itemServiceImpl;
+    private final ItemService itemService;
 
     //상품 전체 조회
     @GetMapping
     public ResponseEntity<List<ItemListResponse>> findAll(){
         return ResponseEntity.ok()
-                .body(itemServiceImpl.findAll());
+                .body(itemService.findAll());
     }
 
     //상품 상세 조회
     @GetMapping("/{itemId}")
     public ResponseEntity<ItemResponse> detailFind(@PathVariable("itemId") long itemId){
         return ResponseEntity.ok()
-                .body(itemServiceImpl.findItemDetailInfo(itemId));
+                .body(itemService.findItemDetailInfo(itemId));
     }
 
 }
