+++ src/test/java/com/project/shop/item/builder/ItemFixture.java
@@ -7,9 +7,21 @@ import com.project.shop.item.dto.request.*;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public class ItemBuilder {
+public class ItemFixture {
 
     static LocalDateTime now = LocalDateTime.now();
+
+    public static Item createItem(Category category,String itemName, int price, String explain, LocalDateTime now){
+
+        return Item.builder()
+                .category(category)
+                .itemName(itemName)
+                .price(price)
+                .explain(explain)
+                .insertDate(now)
+                .updateDate(now)
+                .build();
+    }
 
     public static Item createItem1(Category category){
 
@@ -103,7 +115,7 @@ public class ItemBuilder {
 
     public static ItemRequest createItemRequest1(){
         return new ItemRequest(
-                CategoryBuilder.createCategoryRequest2(),
+                CategoryFixture.createCategoryRequest2(),
                 "조던1",
                 5000,
                 "재고 없음",
@@ -115,7 +127,7 @@ public class ItemBuilder {
     public static ItemRequest createItemRequest2(){
 
         return new ItemRequest(
-                CategoryBuilder.createCategoryRequest2(),
+                CategoryFixture.createCategoryRequest2(),
                 "삼바",
                 109000,
                 "품절이 잦습니다.",
@@ -127,7 +139,7 @@ public class ItemBuilder {
     public static ItemUpdateRequest createItemUpdateRequest(){
 
         return new ItemUpdateRequest(
-                CategoryBuilder.createCategoryUpdateRequest(),
+                CategoryFixture.createCategoryUpdateRequest(),
                 "물",
                 230000,
                 "신상품",
